+++ internal/bloblang/query/functions.go
@@ -447,12 +447,16 @@ func jsonFunction(args ...interface{}) (Function, error) {
 var _ = RegisterFunction(
 	NewFunctionSpec(
 		FunctionCategoryMessage, "meta",
-		"Returns the value of a metadata key from the input message. Since values are extracted from the read-only input message they do NOT reflect changes made from within the map. In order to query metadata mutations made within a mapping use the [`metadata` function](#metadata). This function supports extracting metadata from other messages of a batch with the `from` method.",
+		"Returns the value of a metadata key from the input message. Since values are extracted from the read-only input message they do NOT reflect changes made from within the map. In order to query metadata mutations made within a mapping use the [`root_meta` function](#root_meta). This function supports extracting metadata from other messages of a batch with the `from` method.",
 		NewExampleSpec("",
 			`root.topic = meta("kafka_topic")`,
 		),
 		NewExampleSpec(
-			"The parameter is optional and if omitted the entire metadata contents are returned as a JSON object.",
+			"If the target key does not exist an error is thrown, allowing you to use coalesce or catch methods to fallback to other queries.",
+			`root.topic = meta("nope") | meta("also nope") | "default"`,
+		),
+		NewExampleSpec(
+			"The parameter is optional and if omitted the entire metadata contents are returned as an object.",
 			`root.all_metadata = meta()`,
 		),
 	),
@@ -502,77 +506,32 @@ var _ = RegisterFunction(
 
 var _ = RegisterFunction(
 	NewFunctionSpec(
-		FunctionCategoryMessage, "metadata",
-		"Returns the value of a metadata key, or an empty string if the metadata key is not found. Changes made to metadata during the mapping will be reflected by this function.",
+		FunctionCategoryMessage, "root_meta",
+		"Returns the value of a metadata key from the new message being created. Changes made to metadata during a mapping will be reflected by this function.",
 		NewExampleSpec("",
-			`root.topic = metadata("kafka_topic")`,
+			`root.topic = root_meta("kafka_topic")`,
 		),
 		NewExampleSpec(
-			"The parameter is optional and if omitted the entire metadata contents are returned as a JSON object.",
-			`root.all_metadata = metadata()`,
+			"If the target key does not exist an error is thrown, allowing you to use coalesce or catch methods to fallback to other queries.",
+			`root.topic = root_meta("nope") | root_meta("also nope") | "default"`,
+		),
+		NewExampleSpec(
+			"The parameter is optional and if omitted the entire metadata contents are returned as an object.",
+			`root.all_metadata = root_meta()`,
 		),
 	).Beta(),
 	true,
 	func(args ...interface{}) (Function, error) {
 		if len(args) > 0 {
 			field := args[0].(string)
-			return ClosureFunction("metadata field "+field, func(ctx FunctionContext) (interface{}, error) {
+			return ClosureFunction("root_meta field "+field, func(ctx FunctionContext) (interface{}, error) {
 				if ctx.NewMeta == nil {
-					return nil, errors.New("metadata cannot be queried in this context")
+					return nil, errors.New("root metadata cannot be queried in this context")
 				}
-				return ctx.NewMeta.Get(field), nil
-			}, func(ctx TargetsContext) (TargetsContext, []TargetPath) {
-				paths := []TargetPath{
-					NewTargetPath(TargetMetadata, field),
+				v := ctx.NewMeta.Get(field)
+				if v == "" {
+					return nil, fmt.Errorf("metadata value '%v' not found", field)
 				}
-				ctx = ctx.WithValues(paths)
-				return ctx, paths
-			}), nil
-		}
-		return ClosureFunction("metadata object", func(ctx FunctionContext) (interface{}, error) {
-			if ctx.NewMeta == nil {
-				return nil, errors.New("metadata cannot be queried in this context")
-			}
-			kvs := map[string]interface{}{}
-			ctx.NewMeta.Iter(func(k, v string) error {
-				if len(v) > 0 {
-					kvs[k] = v
-				}
-				return nil
-			})
-			return kvs, nil
-		}, func(ctx TargetsContext) (TargetsContext, []TargetPath) {
-			paths := []TargetPath{
-				NewTargetPath(TargetMetadata),
-			}
-			ctx = ctx.WithValues(paths)
-			return ctx, paths
-		}), nil
-	},
-	ExpectOneOrZeroArgs(),
-	ExpectStringArg(0),
-)
-
-//------------------------------------------------------------------------------
-
-var _ = RegisterFunction(
-	NewFunctionSpec(
-		FunctionCategoryMessage, "source_metadata",
-		"Returns the value of a metadata key from the original input message, or an empty string if the key is not found. This function does NOT reflect changes made from within the map. In order to query metadata mutations made within a mapping use the [`metadata` function](#metadata). This function supports extracting metadata from other messages of a batch with the `from` method.",
-		NewExampleSpec("",
-			`root.topic = source_metadata("kafka_topic")`,
-		),
-		NewExampleSpec(
-			"The parameter is optional and if omitted the entire metadata contents are returned as a JSON object.",
-			`root.all_metadata = source_metadata()`,
-		),
-	).Beta(),
-	true,
-	func(args ...interface{}) (Function, error) {
-		if len(args) > 0 {
-			field := args[0].(string)
-			return ClosureFunction("source_metadata field "+field, func(ctx FunctionContext) (interface{}, error) {
-				v := ctx.MsgBatch.Get(ctx.Index).Metadata().Get(field)
 				return v, nil
 			}, func(ctx TargetsContext) (TargetsContext, []TargetPath) {
 				paths := []TargetPath{
@@ -582,9 +541,12 @@ var _ = RegisterFunction(
 				return ctx, paths
 			}), nil
 		}
-		return ClosureFunction("source_metadata object", func(ctx FunctionContext) (interface{}, error) {
+		return ClosureFunction("root_meta object", func(ctx FunctionContext) (interface{}, error) {
+			if ctx.NewMeta == nil {
+				return nil, errors.New("root metadata cannot be queried in this context")
+			}
 			kvs := map[string]interface{}{}
-			ctx.MsgBatch.Get(ctx.Index).Metadata().Iter(func(k, v string) error {
+			ctx.NewMeta.Iter(func(k, v string) error {
 				if len(v) > 0 {
 					kvs[k] = v
 				}